[issue-244-refactor-network-handling-code] moved all classes to messages/ folder, cleanup
