+++ src/main.cpp
@@ -40,7 +40,7 @@
 #include "undo.h"
 #include "utilmoneystr.h"
 #include "validationinterface.h"
-
+#include "messages/messages.h"
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/filesystem.hpp>
 #include <boost/math/distributions/poisson.hpp>
@@ -90,102 +90,7 @@ std::set<int> setDirtyFileInfo;
 // Internal stuff
 namespace {
 
-    struct CBlockIndexWorkComparator
-    {
-        bool operator()(CBlockIndex *pa, CBlockIndex *pb) const {
-            // First sort by most total work, ...
-            if (pa->nChainWork > pb->nChainWork) return false;
-            if (pa->nChainWork < pb->nChainWork) return true;
 
-            // ... then by earliest time received, ...
-            if (pa->nSequenceId < pb->nSequenceId) return false;
-            if (pa->nSequenceId > pb->nSequenceId) return true;
-
-            // Use pointer address as tie breaker (should only happen with blocks
-            // loaded from disk, as those all have id 0).
-            if (pa < pb) return false;
-            if (pa > pb) return true;
-
-            // Identical blocks.
-            return false;
-        }
-    };
-
-    CBlockIndex *pindexBestInvalid;
-
-    /**
-     * The set of all CBlockIndex entries with BLOCK_VALID_TRANSACTIONS (for itself and all ancestors) and
-     * as good as our current tip or better. Entries may be failed, though, and pruning nodes may be
-     * missing the data for the block.
-     */
-    std::set<CBlockIndex*, CBlockIndexWorkComparator> setBlockIndexCandidates;
-    /** Number of nodes with fSyncStarted. */
-    int nSyncStarted = 0;
-    /** All pairs A->B, where A (or one of its ancestors) misses transactions, but B has transactions.
-     * Pruned nodes may have entries where B is missing data.
-     */
-    std::multimap<CBlockIndex*, CBlockIndex*> mapBlocksUnlinked;
-
-    /** Global flag to indicate we should check to see if there are
-     *  block/undo files that should be deleted.  Set on startup
-     *  or if we allocate more file space when we're in prune mode
-     */
-    bool fCheckForPruning = false;
-
-    /**
-     * Every received block is assigned a unique and increasing identifier, so we
-     * know which one to give priority in case of a fork.
-     */
-    CCriticalSection cs_nBlockSequenceId;
-    /** Blocks loaded from disk are assigned id 0, so start the counter at 1. */
-    uint32_t nBlockSequenceId = 1;
-
-    /**
-     * Sources of received blocks, saved to be able to send them reject
-     * messages or ban them when processing happens afterwards. Protected by
-     * cs_main.
-     */
-    std::map<uint256, NodeId> mapBlockSource;
-
-    /**
-     * Filter for transactions that were recently rejected by
-     * AcceptToMemoryPool. These are not rerequested until the chain tip
-     * changes, at which point the entire filter is reset. Protected by
-     * cs_main.
-     *
-     * Without this filter we'd be re-requesting txs from each of our peers,
-     * increasing bandwidth consumption considerably. For instance, with 100
-     * peers, half of which relay a tx we don't accept, that might be a 50x
-     * bandwidth increase. A flooding attacker attempting to roll-over the
-     * filter using minimum-sized, 60byte, transactions might manage to send
-     * 1000/sec if we have fast peers, so we pick 120,000 to give our peers a
-     * two minute window to send invs to us.
-     *
-     * Decreasing the false positive rate is fairly cheap, so we pick one in a
-     * million to make it highly unlikely for users to have issues with this
-     * filter.
-     *
-     * Memory used: 1.7MB
-     */
-    boost::scoped_ptr<CRollingBloomFilter> recentRejects;
-    uint256 hashRecentRejectsChainTip;
-
-    /** Blocks that are in flight, and that are in the queue to be downloaded. Protected by cs_main. */
-    struct QueuedBlock {
-        uint256 hash;
-        CBlockIndex* pindex;     //!< Optional.
-        bool fValidatedHeaders;  //!< Whether this block has validated headers at the time of request.
-    };
-    std::map<uint256, std::pair<NodeId, std::list<QueuedBlock>::iterator> > mapBlocksInFlight;
-
-    /** Number of preferable block download peers. */
-    int nPreferredDownload = 0;
-
-    /** Dirty block index entries. */
-    std::set<CBlockIndex*> setDirtyBlockIndex;
-
-    /** Number of peers from which we're downloading blocks. */
-    int nPeersWithValidatedDownloads = 0;
 } // anon namespace
 
 //////////////////////////////////////////////////////////////////////////////
@@ -194,69 +99,6 @@ namespace {
 //
 
 namespace {
-
-struct CBlockReject {
-    unsigned char chRejectCode;
-    std::string strRejectReason;
-    uint256 hashBlock;
-};
-
-/**
- * Maintain validation-specific state about nodes, protected by cs_main, instead
- * by CNode's own locks. This simplifies asynchronous operation, where
- * processing of incoming data is done after the ProcessMessage call returns,
- * and we're no longer holding the node's locks.
- */
-struct CNodeState {
-    //! The peer's address
-    CService address;
-    //! Whether we have a fully established connection.
-    bool fCurrentlyConnected;
-    //! Accumulated misbehaviour score for this peer.
-    int nMisbehavior;
-    //! Whether this peer should be disconnected and banned (unless whitelisted).
-    bool fShouldBan;
-    //! List of asynchronously-determined block rejections to notify this peer about.
-    std::vector<CBlockReject> rejects;
-    //! The best known block we know this peer has announced.
-    CBlockIndex *pindexBestKnownBlock;
-    //! The hash of the last unknown block this peer has announced.
-    uint256 hashLastUnknownBlock;
-    //! The last full block we both have.
-    CBlockIndex *pindexLastCommonBlock;
-    //! The best header we have sent our peer.
-    CBlockIndex *pindexBestHeaderSent;
-    //! Whether we've started headers synchronization with this peer.
-    bool fSyncStarted;
-    //! Since when we're stalling block download progress (in microseconds), or 0.
-    int64_t nStallingSince;
-    std::list<QueuedBlock> vBlocksInFlight;
-    //! When the first entry in vBlocksInFlight started downloading. Don't care when vBlocksInFlight is empty.
-    int64_t nDownloadingSince;
-    int nBlocksInFlight;
-    int nBlocksInFlightValidHeaders;
-    //! Whether we consider this a preferred download peer.
-    bool fPreferredDownload;
-    //! Whether this peer wants invs or headers (when possible) for block announcements.
-    bool fPreferHeaders;
-
-    CNodeState() {
-        fCurrentlyConnected = false;
-        nMisbehavior = 0;
-        fShouldBan = false;
-        pindexBestKnownBlock = NULL;
-        hashLastUnknownBlock.SetNull();
-        pindexLastCommonBlock = NULL;
-        pindexBestHeaderSent = NULL;
-        fSyncStarted = false;
-        nStallingSince = 0;
-        nDownloadingSince = 0;
-        nBlocksInFlight = 0;
-        nBlocksInFlightValidHeaders = 0;
-        fPreferredDownload = false;
-        fPreferHeaders = false;
-    }
-};
 
 /** Map maintaining per-node state. Requires cs_main. */
 std::map<NodeId, CNodeState> mapNodeState;
@@ -273,16 +115,6 @@ int GetHeight()
 {
     LOCK(cs_main);
     return chainActive.Height();
-}
-
-void UpdatePreferredDownload(CNode* node, CNodeState* state)
-{
-    nPreferredDownload -= state->fPreferredDownload;
-
-    // Whether this node should be marked as a preferred download node.
-    state->fPreferredDownload = (!node->fInbound || node->fWhitelisted) && !node->fOneShot && !node->fClient;
-
-    nPreferredDownload += state->fPreferredDownload;
 }
 
 void InitializeNode(NodeId nodeid, const CNode *pnode) {
@@ -366,22 +198,8 @@ void MarkBlockAsInFlight(NodeId nodeid, const uint256& hash, const Consensus::Pa
     mapBlocksInFlight[hash] = std::make_pair(nodeid, it);
 }
 
-/** Check whether the last unknown block a peer advertised is not yet known. */
-void ProcessBlockAvailability(NodeId nodeid) {
-    CNodeState *state = State(nodeid);
-    assert(state != NULL);
-
-    if (!state->hashLastUnknownBlock.IsNull()) {
-        auto itOld = Blocks::indexMap.find(state->hashLastUnknownBlock);
-        if (itOld != Blocks::indexMap.end() && itOld->second->nChainWork > 0) {
-            if (state->pindexBestKnownBlock == NULL || itOld->second->nChainWork >= state->pindexBestKnownBlock->nChainWork)
-                state->pindexBestKnownBlock = itOld->second;
-            state->hashLastUnknownBlock.SetNull();
-        }
-    }
-}
-
 /** Update tracking information about which blocks a peer is assumed to have. */
+// Used in HeadersMessage, InvMessage
 void UpdateBlockAvailability(NodeId nodeid, const uint256 &hash) {
     CNodeState *state = State(nodeid);
     assert(state != NULL);
@@ -391,7 +209,8 @@ void UpdateBlockAvailability(NodeId nodeid, const uint256 &hash) {
     auto it = Blocks::indexMap.find(hash);
     if (it != Blocks::indexMap.end() && it->second->nChainWork > 0) {
         // An actually better block was announced.
-        if (state->pindexBestKnownBlock == NULL || it->second->nChainWork >= state->pindexBestKnownBlock->nChainWork)
+        if (state->pindexBestKnownBlock == NULL ||
+            it->second->nChainWork >= state->pindexBestKnownBlock->nChainWork)
             state->pindexBestKnownBlock = it->second;
     } else {
         // An unknown block was announced; just assume that the latest one is the best one.
@@ -400,6 +219,7 @@ void UpdateBlockAvailability(NodeId nodeid, const uint256 &hash) {
 }
 
 // Requires cs_main
+// Used by HeadersMessage and InvMessage
 bool CanDirectFetch(const Consensus::Params &consensusParams)
 {
     return chainActive.Tip()->GetBlockTime() > GetAdjustedTime() - consensusParams.nPowTargetSpacing * 20;
@@ -1214,11 +1034,6 @@ bool GetTransaction(const uint256 &hash, CTransaction &txOut, const Consensus::P
 
     return false;
 }
-
-
-
-
-
 
 //////////////////////////////////////////////////////////////////////////////
 //
@@ -2146,13 +1961,6 @@ bool ConnectBlock(const CBlock& block, CValidationState& state, CBlockIndex* pin
 
     return true;
 }
-
-enum FlushStateMode {
-    FLUSH_STATE_NONE,
-    FLUSH_STATE_IF_NEEDED,
-    FLUSH_STATE_PERIODIC,
-    FLUSH_STATE_ALWAYS
-};
 
 /**
  * Update the on-disk chain state.
@@ -3127,52 +2935,6 @@ bool ContextualCheckBlock(const CBlock& block, CValidationState& state, CBlockIn
     return true;
 }
 
-static bool AcceptBlockHeader(const CBlockHeader& block, CValidationState& state, const CChainParams& chainparams, CBlockIndex** ppindex=NULL)
-{
-    AssertLockHeld(cs_main);
-    // Check for duplicate
-    uint256 hash = block.GetHash();
-    auto miSelf = Blocks::indexMap.find(hash);
-    CBlockIndex *pindex = NULL;
-    if (hash != chainparams.GetConsensus().hashGenesisBlock) {
-
-        if (miSelf != Blocks::indexMap.end()) {
-            // Block header is already known.
-            pindex = miSelf->second;
-            if (ppindex)
-                *ppindex = pindex;
-            if (pindex->nStatus & BLOCK_FAILED_MASK)
-                return state.Invalid(error("%s: block is marked invalid", __func__), 0, "duplicate");
-            return true;
-        }
-
-        if (!CheckBlockHeader(block, state))
-            return false;
-
-        // Get prev block index
-        CBlockIndex* pindexPrev = NULL;
-        auto mi = Blocks::indexMap.find(block.hashPrevBlock);
-        if (mi == Blocks::indexMap.end())
-            return state.DoS(10, error("%s: prev block not found", __func__), 0, "bad-prevblk");
-        pindexPrev = (*mi).second;
-        if (pindexPrev->nStatus & BLOCK_FAILED_MASK)
-            return state.DoS(100, error("%s: prev block invalid", __func__), REJECT_INVALID, "bad-prevblk");
-
-        assert(pindexPrev);
-        if (fCheckpointsEnabled && !CheckIndexAgainstCheckpoint(pindexPrev, state, chainparams, hash))
-            return error("%s: CheckIndexAgainstCheckpoint(): %s", __func__, state.GetRejectReason().c_str());
-
-        if (!ContextualCheckBlockHeader(block, state, pindexPrev))
-            return false;
-    }
-    if (pindex == NULL)
-        pindex = AddToBlockIndex(block);
-
-    if (ppindex)
-        *ppindex = pindex;
-
-    return true;
-}
 
 /** Store block on disk. If dbp is non-NULL, the file is known to already reside on disk */
 static bool AcceptBlock(const CBlock& block, CValidationState& state, const CChainParams& chainparams, CBlockIndex** ppindex, bool fRequested, CDiskBlockPos* dbp)
@@ -3960,1761 +3722,6 @@ std::string GetWarnings(const std::string& strFor)
     return "error";
 }
 
-
-
-
-
-
-
-
-//////////////////////////////////////////////////////////////////////////////
-//
-// Messages
-//
-
-
-bool static AlreadyHave(const CInv& inv) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    switch (inv.type)
-    {
-    case MSG_TX:
-        {
-            assert(recentRejects);
-            if (chainActive.Tip()->GetBlockHash() != hashRecentRejectsChainTip)
-            {
-                // If the chain tip has changed previously rejected transactions
-                // might be now valid, e.g. due to a nLockTime'd tx becoming valid,
-                // or a double-spend. Reset the rejects filter and give those
-                // txs a second chance.
-                hashRecentRejectsChainTip = chainActive.Tip()->GetBlockHash();
-                recentRejects->reset();
-            }
-
-            return recentRejects->contains(inv.hash) ||
-                   mempool.exists(inv.hash) ||
-                   CTxOrphanCache::contains(inv.hash) ||
-                   pcoinsTip->HaveCoins(inv.hash);
-        }
-    case MSG_BLOCK:
-        return Blocks::indexMap.count(inv.hash);
-    }
-    // Don't know what it is, just say we already got one
-    return true;
-}
-
-void static ProcessGetData(CNode* pfrom, const Consensus::Params& consensusParams)
-{
-    std::deque<CInv>::iterator it = pfrom->vRecvGetData.begin();
-
-    std::vector<CInv> vNotFound;
-
-    LOCK(cs_main);
-
-    while (it != pfrom->vRecvGetData.end()) {
-        // Don't bother if send buffer is too full to respond anyway
-        if (pfrom->nSendSize >= SendBufferSize())
-            break;
-
-        const CInv &inv = *it;
-        {
-            boost::this_thread::interruption_point();
-            it++;
-
-            if (inv.type == MSG_BLOCK || inv.type == MSG_FILTERED_BLOCK
-                    || inv.type == MSG_THINBLOCK || inv.type == MSG_XTHINBLOCK)
-            {
-                bool send = false;
-                auto mi = Blocks::indexMap.find(inv.hash);
-                if (mi != Blocks::indexMap.end())
-                {
-                    if (chainActive.Contains(mi->second)) {
-                        send = true;
-                    } else {
-                        static const int nOneMonth = 30 * 24 * 60 * 60;
-                        // To prevent fingerprinting attacks, only send blocks outside of the active
-                        // chain if they are valid, and no more than a month older (both in time, and in
-                        // best equivalent proof of work) than the best header chain we know about.
-                        send = mi->second->IsValid(BLOCK_VALID_SCRIPTS) && (pindexBestHeader != NULL) &&
-                            (pindexBestHeader->GetBlockTime() - mi->second->GetBlockTime() < nOneMonth) &&
-                            (GetBlockProofEquivalentTime(*pindexBestHeader, *mi->second, *pindexBestHeader, consensusParams) < nOneMonth);
-                        if (!send) {
-                            LogPrintf("%s: ignoring request from peer=%i for old block that isn't in the main chain\n", __func__, pfrom->GetId());
-                        }
-                    }
-                }
-                // disconnect node in case we have reached the outbound limit for serving historical blocks
-                // never disconnect whitelisted nodes
-                static const int nOneWeek = 7 * 24 * 60 * 60; // assume > 1 week = historical
-                if (send && CNode::OutboundTargetReached(true) && ( ((pindexBestHeader != NULL) && (pindexBestHeader->GetBlockTime() - mi->second->GetBlockTime() > nOneWeek)) || inv.type == MSG_FILTERED_BLOCK) && !pfrom->fWhitelisted)
-                {
-                    LogPrint("net", "historical block serving limit reached, disconnect peer=%d\n", pfrom->GetId());
-
-                    //disconnect node
-                    pfrom->fDisconnect = true;
-                    send = false;
-                }
-                // Pruned nodes may have deleted the block, so check whether
-                // it's available before trying to send.
-                if (send && (mi->second->nStatus & BLOCK_HAVE_DATA))
-                {
-                    // Send block from disk
-                    CBlock block;
-                    if (!ReadBlockFromDisk(block, (*mi).second, consensusParams))
-                        assert(!"cannot load block from disk");
-
-                    bool sendFullBlock = true;
-
-                    if (inv.type == MSG_XTHINBLOCK) {
-                        CXThinBlock xThinBlock(block, pfrom->pThinBlockFilter);
-                        if (!xThinBlock.collision) {
-                            const int nSizeBlock = ::GetSerializeSize(block, SER_NETWORK, PROTOCOL_VERSION);
-                            // Only send a thinblock if smaller than a regular block
-                            const int nSizeThinBlock = ::GetSerializeSize(xThinBlock, SER_NETWORK, PROTOCOL_VERSION);
-                            if (nSizeThinBlock < nSizeBlock) {
-                                pfrom->PushMessage(NetMsgType::XTHINBLOCK, xThinBlock);
-                                sendFullBlock = false;
-                                LogPrint("thin", "Sent xthinblock - size: %d vs block size: %d => tx hashes: %d transactions: %d  peerid=%d\n",
-                                         nSizeThinBlock, nSizeBlock, xThinBlock.vTxHashes.size(), xThinBlock.vMissingTx.size(), pfrom->id);
-                            }
-                        }
-                    }
-                    else if (inv.type == MSG_FILTERED_BLOCK)
-                    {
-                        LOCK(pfrom->cs_filter);
-                        if (pfrom->pfilter)
-                        {
-                            CMerkleBlock merkleBlock(block, *pfrom->pfilter);
-                            pfrom->PushMessage(NetMsgType::MERKLEBLOCK, merkleBlock);
-                            // CMerkleBlock just contains hashes, so also push any transactions in the block the client did not see
-                            // This avoids hurting performance by pointlessly requiring a round-trip
-                            // Note that there is currently no way for a node to request any single transactions we didn't send here -
-                            // they must either disconnect and retry or request the full block.
-                            // Thus, the protocol spec specified allows for us to provide duplicate txn here,
-                            // however we MUST always provide at least what the remote peer needs
-                            typedef std::pair<unsigned int, uint256> PairType;
-                            BOOST_FOREACH(PairType& pair, merkleBlock.vMatchedTxn)
-                                pfrom->PushMessage(NetMsgType::TX, block.vtx[pair.first]);
-                            sendFullBlock = false;
-                        }
-                    }
-                    if (sendFullBlock) // if none of the other methods were actually executed;
-                         pfrom->PushMessage(NetMsgType::BLOCK, block);
-
-                    // Trigger the peer node to send a getblocks request for the next batch of inventory
-                    if (inv.hash == pfrom->hashContinue)
-                    {
-                        // Bypass PushInventory, this must send even if redundant,
-                        // and we want it right after the last block so they don't
-                        // wait for other stuff first.
-                        std::vector<CInv> vInv;
-                        vInv.push_back(CInv(MSG_BLOCK, chainActive.Tip()->GetBlockHash()));
-                        pfrom->PushMessage(NetMsgType::INV, vInv);
-                        pfrom->hashContinue.SetNull();
-                    }
-                }
-            }
-            else if (inv.IsKnownType())
-            {
-                // Send stream from relay memory
-                bool pushed = false;
-                {
-                    LOCK(cs_mapRelay);
-                    std::map<CInv, CDataStream>::iterator mi = mapRelay.find(inv);
-                    if (mi != mapRelay.end()) {
-                        pfrom->PushMessage(inv.GetCommand(), (*mi).second);
-                    }
-                }
-                if (!pushed && inv.type == MSG_TX) {
-                    CTransaction tx;
-                    if (mempool.lookup(inv.hash, tx)) {
-                        CDataStream ss(SER_NETWORK, PROTOCOL_VERSION);
-                        ss.reserve(1000);
-                        ss << tx;
-                        pfrom->PushMessage(NetMsgType::TX, ss);
-                        pushed = true;
-                    }
-                }
-                if (!pushed) {
-                    vNotFound.push_back(inv);
-                }
-            }
-
-            // Track requests for our stuff.
-            GetMainSignals().Inventory(inv.hash);
-
-            if (inv.type == MSG_BLOCK || inv.type == MSG_FILTERED_BLOCK
-                    || inv.type == MSG_THINBLOCK || inv.type == MSG_XTHINBLOCK)
-                break;
-        }
-    }
-
-    pfrom->vRecvGetData.erase(pfrom->vRecvGetData.begin(), it);
-
-    if (!vNotFound.empty()) {
-        // Let the peer know that we didn't find what it asked for, so it doesn't
-        // have to wait around forever. Currently only SPV clients actually care
-        // about this message: it's needed when they are recursively walking the
-        // dependencies of relevant unconfirmed transactions. SPV clients want to
-        // do that because they want to know about (and store and rebroadcast and
-        // risk analyze) the dependencies of transactions relevant to them, without
-        // having to download the entire memory pool.
-        pfrom->PushMessage(NetMsgType::NOTFOUND, vNotFound);
-    }
-}
-
-class NetworkMessage {
-public:
-    virtual ~NetworkMessage() {}
-
-    NetworkMessage() {}
-
-    virtual bool handle(CNode * const pfrom,
-                        CDataStream &vRecv,
-                        int64_t nTimeReceived,
-                        std::string &strCommand,
-                        const CChainParams &chainparams,
-                        const bool xthinEnabled,
-                        const bool fReindexing) = 0;
-};
-
-class VersionMessage : public NetworkMessage {
-public:
-    ~VersionMessage() {}
-
-    VersionMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chain_params,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        // Each connection can only send one version message
-        if (pfrom->nVersion != 0) {
-            pfrom->PushMessage(NetMsgType::REJECT, strCommand, REJECT_DUPLICATE,
-                               std::string("Duplicate version message"));
-            Misbehaving(pfrom->GetId(), 1);
-            return false;
-        }
-
-        int64_t nTime;
-        CAddress addrMe;
-        CAddress addrFrom;
-        uint64_t nNonce = 1;
-        vRecv >> pfrom->nVersion >> pfrom->nServices >> nTime >> addrMe;
-
-        if (pfrom->nVersion < MIN_PEER_PROTO_VERSION) {
-            // disconnect from peers older than this proto version
-            LogPrintf("peer=%d using obsolete version %i; disconnecting\n", pfrom->id, pfrom->nVersion);
-            pfrom->PushMessage(NetMsgType::REJECT, strCommand, REJECT_OBSOLETE,
-                               strprintf("Version must be %d or greater", MIN_PEER_PROTO_VERSION));
-            pfrom->fDisconnect = true;
-            addrman.increaseUselessness(pfrom->addr, 2);
-            return false;
-        }
-
-        if (pfrom->nVersion == 10300)
-            pfrom->nVersion = 300;
-        if (!vRecv.empty())
-            vRecv >> addrFrom >> nNonce;
-        if (!vRecv.empty()) {
-            vRecv >> LIMITED_STRING(pfrom->strSubVer, MAX_SUBVERSION_LENGTH);
-            pfrom->cleanSubVer = SanitizeString(pfrom->strSubVer);
-        }
-        if (!vRecv.empty())
-            vRecv >> pfrom->nStartingHeight;
-        if (!vRecv.empty())
-            vRecv >> pfrom->fRelayTxes; // set to true after we get the first filter* message
-        else
-            pfrom->fRelayTxes = true;
-
-        // Disconnect if we connected to ourself
-        if (nNonce == nLocalHostNonce && nNonce > 1) {
-            LogPrintf("connected to self at %s, disconnecting\n", pfrom->addr.ToString());
-            pfrom->fDisconnect = true;
-            return true;
-        }
-
-        pfrom->addrLocal = addrMe;
-        if (pfrom->fInbound && addrMe.IsRoutable()) {
-            SeenLocal(addrMe);
-        }
-
-        // Be shy and don't send version until we hear
-        if (pfrom->fInbound)
-            pfrom->PushVersion();
-
-        pfrom->fClient = !(pfrom->nServices & NODE_NETWORK);
-
-        // Potentially mark this peer as a preferred download peer.
-        UpdatePreferredDownload(pfrom, State(pfrom->GetId()));
-
-        // Change version
-        pfrom->PushMessage(NetMsgType::VERACK);
-        pfrom->ssSend.SetVersion(std::min(pfrom->nVersion, PROTOCOL_VERSION));
-
-        if (!pfrom->fInbound) {
-            // Advertise our address
-            if (fListen && !IsInitialBlockDownload()) {
-                CAddress addr = GetLocalAddress(&pfrom->addr);
-                if (addr.IsRoutable()) {
-                    LogPrintf("ProcessMessages: advertising address %s\n", addr.ToString());
-                    pfrom->PushAddress(addr);
-                } else if (IsPeerAddrLocalGood(pfrom)) {
-                    addr.SetIP(pfrom->addrLocal);
-                    LogPrintf("ProcessMessages: advertising address %s\n", addr.ToString());
-                    pfrom->PushAddress(addr);
-                }
-            }
-
-            // Get recent addresses
-            if (pfrom->fOneShot || pfrom->nVersion >= CADDR_TIME_VERSION || addrman.size() < 1000) {
-                pfrom->PushMessage(NetMsgType::GETADDR);
-                pfrom->fGetAddr = true;
-            }
-            addrman.Good(pfrom->addr);
-        } else {
-            if (((CNetAddr) pfrom->addr) == (CNetAddr) addrFrom) {
-                addrman.Add(addrFrom, addrFrom);
-                addrman.Good(addrFrom);
-            }
-        }
-
-        pfrom->fSuccessfullyConnected = true;
-        logCritical(Log::Net) << "receive version message:" << pfrom->addr << pfrom->cleanSubVer << "version:"
-                              << pfrom->nVersion << "blocks:" << pfrom->nStartingHeight << "id:" << pfrom->id;
-        if (pfrom->isCashNode)
-            logInfo(Log::Net) << "peer id:" << pfrom->GetId() << "uses CASH message-headers";
-
-        int64_t nTimeOffset = nTime - GetTime();
-        pfrom->nTimeOffset = nTimeOffset;
-        AddTimeData(pfrom->addr, nTimeOffset);
-
-        return true;
-    }
-};
-
-class VerAckMessage : public NetworkMessage {
-public:
-    ~VerAckMessage() {}
-
-    VerAckMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        pfrom->SetRecvVersion(std::min(pfrom->nVersion, PROTOCOL_VERSION));
-
-        // Mark this node as currently connected, so we update its timestamp later.
-        if (pfrom->fNetworkNode) {
-            LOCK(cs_main);
-            State(pfrom->GetId())->fCurrentlyConnected = true;
-        }
-
-        if (pfrom->nVersion >= SENDHEADERS_VERSION) {
-            // Tell our peer we prefer to receive headers rather than inv's
-            // We send this to non-NODE NETWORK peers as well, because even
-            // non-NODE NETWORK peers can announce blocks (such as pruning
-            // nodes)
-
-            // BUIP010 Extreme Thinblocks: We only do inv/getdata for xthinblocks and so we must have headersfirst turned off
-            if (!xthinEnabled)
-                pfrom->PushMessage(NetMsgType::SENDHEADERS);
-        }
-
-        // send our listening port in a separate version message
-        if (pfrom->nVersion >= EXPEDITED_VERSION)
-            pfrom->PushMessage(NetMsgType::VERSION2, GetListenPort());
-
-        return true;
-    }
-};
-
-class AddrMessage : public NetworkMessage {
-public:
-    ~AddrMessage() {}
-
-    AddrMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        std::vector<CAddress> vAddr;
-        vRecv >> vAddr;
-
-        // Don't want addr from older versions unless seeding
-        if (pfrom->nVersion < CADDR_TIME_VERSION && addrman.size() > 1000)
-            return true;
-        if (vAddr.size() > 1000) {
-            Misbehaving(pfrom->GetId(), 20);
-            return error("message addr size() = %u", vAddr.size());
-        }
-
-        // Store the new addresses
-        std::vector<CAddress> vAddrOk;
-        int64_t nNow = GetAdjustedTime();
-        int64_t nSince = nNow - 10 * 60;
-        BOOST_FOREACH(CAddress &addr, vAddr) {
-            boost::this_thread::interruption_point();
-
-            if (addr.nTime <= 100000000 || addr.nTime > nNow + 10 * 60)
-                addr.nTime = nNow - 5 * 24 * 60 * 60;
-            pfrom->AddAddressKnown(addr);
-            bool fReachable = IsReachable(addr);
-            if (addr.nTime > nSince && !pfrom->fGetAddr && vAddr.size() <= 10 && addr.IsRoutable()) {
-                // Relay to a limited number of other nodes
-                {
-                    LOCK(cs_vNodes);
-                    // Use deterministic randomness to send to the same nodes for 24 hours
-                    // at a time so the addrKnowns of the chosen nodes prevent repeats
-                    static uint256 hashSalt;
-                    if (hashSalt.IsNull())
-                        hashSalt = GetRandHash();
-                    uint64_t hashAddr = addr.GetHash();
-                    uint256 hashRand = ArithToUint256(UintToArith256(hashSalt) ^ (hashAddr << 32) ^
-                                                      ((GetTime() + hashAddr) / (24 * 60 * 60)));
-                    hashRand = Hash(BEGIN(hashRand), END(hashRand));
-                    std::multimap<uint256, CNode *> mapMix;
-                    BOOST_FOREACH(CNode *pnode, vNodes) {
-                        if (pnode->nVersion < CADDR_TIME_VERSION)
-                            continue;
-                        unsigned int nPointer;
-                        memcpy(&nPointer, &pnode, sizeof(nPointer));
-                        uint256 hashKey = ArithToUint256(UintToArith256(hashRand) ^ nPointer);
-                        hashKey = Hash(BEGIN(hashKey), END(hashKey));
-                        mapMix.insert(std::make_pair(hashKey, pnode));
-                    }
-                    int nRelayNodes = fReachable ? 2
-                                                 : 1; // limited relaying of addresses outside our network(s)
-                    for (std::multimap<uint256, CNode *>::iterator mi = mapMix.begin();
-                         mi != mapMix.end() && nRelayNodes-- > 0; ++mi)
-                        ((*mi).second)->PushAddress(addr);
-                }
-            }
-            // Do not store addresses outside our network
-            if (fReachable)
-                vAddrOk.push_back(addr);
-        }
-        addrman.Add(vAddrOk, pfrom->addr, 2 * 60 * 60);
-        if (vAddr.size() < 1000)
-            pfrom->fGetAddr = false;
-        if (pfrom->fOneShot)
-            pfrom->fDisconnect = true;
-        return true;
-    }
-};
-
-class SendHeadersMessage : public NetworkMessage {
-public:
-    ~SendHeadersMessage() {}
-
-    SendHeadersMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        LOCK(cs_main);
-        // BUIP010 Xtreme Thinblocks: We only do inv/getdata for xthinblocks and so we must have headersfirst turned off
-        if (xthinEnabled)
-            State(pfrom->GetId())->fPreferHeaders = false;
-        else
-            State(pfrom->GetId())->fPreferHeaders = true;
-        return true;
-    }
-};
-
-class InvMessage : public NetworkMessage {
-public:
-    ~InvMessage() {}
-
-    InvMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (Application::uahfChainState() == Application::UAHFWaiting) {
-            // this means we are not just in initial block download, we are in a state
-            // where filling the mempool or getting the latest block just doesn't make any sense.
-            // This avoids us banning CASH nodes before we follow the UAHF rules.
-            if (Params().uahfForkBlockHeight() > chainActive.Height())
-                return true;
-        }
-        std::vector<CInv> vInv;
-        vRecv >> vInv;
-        if (vInv.size() > MAX_INV_SZ)
-        {
-            Misbehaving(pfrom->GetId(), 20);
-            return error("message inv size() = %u", vInv.size());
-        }
-
-        bool fBlocksOnly = GetBoolArg("-blocksonly", DEFAULT_BLOCKSONLY);
-
-        // Allow whitelisted peers to send data other than blocks in blocks only mode if whitelistrelay is true
-        if (pfrom->fWhitelisted && GetBoolArg("-whitelistrelay", DEFAULT_WHITELISTRELAY))
-            fBlocksOnly = false;
-
-        LOCK(cs_main);
-
-        std::vector<CInv> vToFetch;
-
-        for (unsigned int nInv = 0; nInv < vInv.size(); nInv++)
-        {
-            const CInv &inv = vInv[nInv];
-
-            boost::this_thread::interruption_point();
-            pfrom->AddInventoryKnown(inv);
-
-            bool fAlreadyHave = AlreadyHave(inv);
-            logDebug(Log::Net) << "got inv:" << inv << (fAlreadyHave ? "have" : "new") << "peer:" << pfrom->id;
-
-            if (inv.type == MSG_BLOCK) {
-                UpdateBlockAvailability(pfrom->GetId(), inv.hash);
-                if (!fAlreadyHave && !fImporting && !fReindexing && !mapBlocksInFlight.count(inv.hash)) {
-                    // First request the headers preceding the announced block. In the normal fully-synced
-                    // case where a new block is announced that succeeds the current tip (no reorganization),
-                    // there are no such headers.
-                    // Secondly, and only when we are close to being synced, we request the announced block directly,
-                    // to avoid an extra round-trip. Note that we must *first* ask for the headers, so by the
-                    // time the block arrives, the header chain leading up to it is already validated. Not
-                    // doing this will result in the received block being rejected as an orphan in case it is
-                    // not a direct successor.
-                    pfrom->PushMessage(NetMsgType::GETHEADERS, chainActive.GetLocator(pindexBestHeader), inv.hash);
-                    CNodeState *nodestate = State(pfrom->GetId());
-                    if (CanDirectFetch(chainparams.GetConsensus()) &&
-                        nodestate->nBlocksInFlight < MAX_BLOCKS_IN_TRANSIT_PER_PEER) {
-                        // BUIP010 Xtreme Thinblocks: begin section
-                        CInv inv2(inv);
-                        CDataStream ss(SER_NETWORK, PROTOCOL_VERSION);
-                        if (xthinEnabled && IsChainNearlySyncd()) {
-                            if (HaveThinblockNodes() && CheckThinblockTimer(inv.hash)) {
-                                // Must download a block from a ThinBlock peer
-                                if (pfrom->mapThinBlocksInFlight.size() < 1 && pfrom->ThinBlockCapable()) { // We can only send one thinblock per peer at a time
-                                    pfrom->mapThinBlocksInFlight[inv2.hash] = GetTime();
-                                    inv2.type = MSG_XTHINBLOCK;
-                                    CBloomFilter filterMemPool = createSeededBloomFilter(CTxOrphanCache::instance()->fetchTransactionIds());
-                                    ss << inv2;
-                                    ss << filterMemPool;
-                                    pfrom->PushMessage(NetMsgType::GET_XTHIN, ss);
-                                    MarkBlockAsInFlight(pfrom->GetId(), inv.hash, chainparams.GetConsensus());
-                                    LogPrint("thin", "Requesting Thinblock %s from peer %s (%d)\n", inv2.hash.ToString(), pfrom->addrName.c_str(),pfrom->id);
-                                }
-                            }
-                            else {
-                                // Try to download a thinblock if possible otherwise just download a regular block
-                                if (pfrom->mapThinBlocksInFlight.size() < 1 && pfrom->ThinBlockCapable()) { // We can only send one thinblock per peer at a time
-                                    pfrom->mapThinBlocksInFlight[inv2.hash] = GetTime();
-                                    inv2.type = MSG_XTHINBLOCK;
-                                    CBloomFilter filterMemPool = createSeededBloomFilter(CTxOrphanCache::instance()->fetchTransactionIds());
-                                    ss << inv2;
-                                    ss << filterMemPool;
-                                    pfrom->PushMessage(NetMsgType::GET_XTHIN, ss);
-                                    LogPrint("thin", "Requesting Thinblock %s from peer %s (%d)\n", inv2.hash.ToString(), pfrom->addrName.c_str(),pfrom->id);
-                                }
-                                else {
-                                    LogPrint("thin", "Requesting Regular Block %s from peer %s (%d)\n", inv2.hash.ToString(), pfrom->addrName.c_str(),pfrom->id);
-                                    vToFetch.push_back(inv2);
-                                }
-                                MarkBlockAsInFlight(pfrom->GetId(), inv.hash, chainparams.GetConsensus());
-                            }
-                        }
-                        else {
-                            vToFetch.push_back(inv2);
-                            MarkBlockAsInFlight(pfrom->GetId(), inv.hash, chainparams.GetConsensus());
-                            LogPrint("thin", "Requesting Regular Block %s from peer %s (%d)\n", inv2.hash.ToString(), pfrom->addrName.c_str(),pfrom->id);
-                        }
-                        // BUIP010 Xtreme Thinblocks: end section
-                    }
-                    LogPrint("net", "getheaders (%d) %s to peer=%d\n", pindexBestHeader->nHeight, inv.hash.ToString(), pfrom->id);
-                }
-            }
-            else
-            {
-                if (fBlocksOnly)
-                    LogPrint("net", "transaction (%s) inv sent in violation of protocol peer=%d\n", inv.hash.ToString(), pfrom->id);
-                else if (!fAlreadyHave && !fImporting && !fReindexing)
-                    pfrom->AskFor(inv);
-            }
-
-            // Track requests for our stuff
-            GetMainSignals().Inventory(inv.hash);
-
-            if (pfrom->nSendSize > (SendBufferSize() * 2)) {
-                Misbehaving(pfrom->GetId(), 50);
-                return error("send buffer size() = %u", pfrom->nSendSize);
-            }
-        }
-
-        if (!vToFetch.empty())
-            pfrom->PushMessage(NetMsgType::GETDATA, vToFetch);
-        return true;
-    }
-};
-
-class GetDataMessage : public NetworkMessage {
-public:
-    ~GetDataMessage(){}
-
-    GetDataMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        std::vector<CInv> vInv;
-        vRecv >> vInv;
-        if (vInv.size() > MAX_INV_SZ)
-        {
-            Misbehaving(pfrom->GetId(), 20);
-            return error("message getdata size() = %u", vInv.size());
-        }
-
-        if (fDebug || (vInv.size() != 1))
-            logDebug(Log::Net) << "received getdata (" << vInv.size() << "invsz) peer:" << pfrom->id;
-
-        if ((fDebug && vInv.size() > 0) || (vInv.size() == 1))
-            logDebug(Log::Net) << "received getdata for:" << vInv[0].ToString() << "peer:" << pfrom->id;
-
-        pfrom->vRecvGetData.insert(pfrom->vRecvGetData.end(), vInv.begin(), vInv.end());
-        ProcessGetData(pfrom, chainparams.GetConsensus());
-        return true;
-    }
-};
-
-class GetBlocksMessage : public NetworkMessage {
-public:
-    ~GetBlocksMessage() {}
-
-    GetBlocksMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        CBlockLocator locator;
-        uint256 hashStop;
-        vRecv >> locator >> hashStop;
-
-        LOCK(cs_main);
-
-        // Find the last block the caller has in the main chain
-        CBlockIndex* pindex = FindForkInGlobalIndex(chainActive, locator);
-
-        // Send the rest of the chain
-        if (pindex)
-            pindex = chainActive.Next(pindex);
-        int nLimit = 500;
-        LogPrint("net", "getblocks %d to %s limit %d from peer=%d\n", (pindex ? pindex->nHeight : -1), hashStop.IsNull() ? "end" : hashStop.ToString(), nLimit, pfrom->id);
-        for (; pindex; pindex = chainActive.Next(pindex))
-        {
-            if (pindex->GetBlockHash() == hashStop)
-            {
-                LogPrint("net", "  getblocks stopping at %d %s\n", pindex->nHeight, pindex->GetBlockHash().ToString());
-                break;
-            }
-            // If pruning, don't inv blocks unless we have on disk and are likely to still have
-            // for some reasonable time window (1 hour) that block relay might require.
-            const int nPrunedBlocksLikelyToHave = MIN_BLOCKS_TO_KEEP - 3600 / chainparams.GetConsensus().nPowTargetSpacing;
-            if (fPruneMode && (!(pindex->nStatus & BLOCK_HAVE_DATA) || pindex->nHeight <= chainActive.Tip()->nHeight - nPrunedBlocksLikelyToHave))
-            {
-                LogPrint("net", " getblocks stopping, pruned or too old block at %d %s\n", pindex->nHeight, pindex->GetBlockHash().ToString());
-                break;
-            }
-            pfrom->PushInventory(CInv(MSG_BLOCK, pindex->GetBlockHash()));
-            if (--nLimit <= 0)
-            {
-                // When this block is requested, we'll send an inv that'll
-                // trigger the peer to getblocks the next batch of inventory.
-                LogPrint("net", "  getblocks stopping at limit %d %s\n", pindex->nHeight, pindex->GetBlockHash().ToString());
-                pfrom->hashContinue = pindex->GetBlockHash();
-                break;
-            }
-        }
-        return true;
-    }
-};
-
-class GetHeadersMessage : public NetworkMessage {
-public:
-    ~GetHeadersMessage() {}
-
-    GetHeadersMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        CBlockLocator locator;
-        uint256 hashStop;
-        vRecv >> locator >> hashStop;
-
-        LOCK(cs_main);
-        if (IsInitialBlockDownload() && !pfrom->fWhitelisted) {
-            logDebug(Log::Net) << "Ignoring getheaders from peer=" <<pfrom->id << "because node is in initial block download";
-            return true;
-        }
-
-        CNodeState *nodestate = State(pfrom->GetId());
-        CBlockIndex* pindex = NULL;
-        if (locator.IsNull())
-        {
-            // If locator is null, return the hashStop block
-            auto mi = Blocks::indexMap.find(hashStop);
-            if (mi == Blocks::indexMap.end())
-                return true;
-            pindex = (*mi).second;
-        }
-        else
-        {
-            // Find the last block the caller has in the main chain
-            pindex = FindForkInGlobalIndex(chainActive, locator);
-            if (pindex)
-                pindex = chainActive.Next(pindex);
-        }
-
-        // we must use CBlocks, as CBlockHeaders won't include the 0x00 nTx count at the end
-        std::vector<CBlock> vHeaders;
-        int nLimit = MAX_HEADERS_RESULTS;
-        LogPrint("net", "getheaders %d to %s from peer=%d\n", (pindex ? pindex->nHeight : -1), hashStop.ToString(), pfrom->id);
-        for (; pindex; pindex = chainActive.Next(pindex))
-        {
-            vHeaders.push_back(pindex->GetBlockHeader());
-            if (--nLimit <= 0 || pindex->GetBlockHash() == hashStop)
-                break;
-        }
-        // pindex can be NULL either if we sent chainActive.Tip() OR
-        // if our peer has chainActive.Tip() (and thus we are sending an empty
-        // headers message). In both cases it's safe to update
-        // pindexBestHeaderSent to be our tip.
-        nodestate->pindexBestHeaderSent = pindex ? pindex : chainActive.Tip();
-        pfrom->PushMessage(NetMsgType::HEADERS, vHeaders);
-
-        return true;
-    }
-};
-
-class TxMessage : public NetworkMessage {
-public:
-    ~TxMessage() {}
-
-    TxMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        // Stop processing the transaction early if
-        // We are in blocks only mode and peer is either not whitelisted or whitelistrelay is off
-        if (GetBoolArg("-blocksonly", DEFAULT_BLOCKSONLY) && (!pfrom->fWhitelisted || !GetBoolArg("-whitelistrelay", DEFAULT_WHITELISTRELAY)))
-        {
-            LogPrint("net", "transaction sent in violation of protocol peer=%d\n", pfrom->id);
-            return true;
-        }
-
-        std::vector<uint256> vWorkQueue;
-        std::vector<uint256> vEraseQueue;
-        CTransaction tx;
-        vRecv >> tx;
-
-        CInv inv(MSG_TX, tx.GetHash());
-        pfrom->AddInventoryKnown(inv);
-
-        LOCK(cs_main);
-
-        bool fMissingInputs = false;
-        CValidationState state;
-
-        pfrom->setAskFor.erase(inv.hash);
-        mapAlreadyAskedFor.erase(inv.hash);
-
-        if (!AlreadyHave(inv) && AcceptToMemoryPool(mempool, state, tx, true, &fMissingInputs))
-        {
-            mempool.check(pcoinsTip);
-            RelayTransaction(tx);
-            vWorkQueue.push_back(inv.hash);
-
-            logDebug(Log::Mempool) << "AcceptToMemoryPool: peer" << pfrom->id << "accepted" << tx.GetHash()
-                                   << "(poolsz" << mempool.size() << "txn," << (mempool.DynamicMemoryUsage() / 1000) << "kB)";
-
-            // Recursively process any orphan transactions that depended on this one
-            std::set<NodeId> setMisbehaving;
-            for (unsigned int i = 0; i < vWorkQueue.size(); i++)
-            {
-                auto orphans = CTxOrphanCache::instance()->fetchTransactionsByPrev(vWorkQueue[i]);
-                for (auto mi = orphans.begin(); mi != orphans.end(); ++mi) {
-                    const CTransaction& orphanTx = mi->tx;
-                    const uint256 orphanHash = orphanTx.GetHash();
-                    bool fMissingInputs2 = false;
-                    // Use a dummy CValidationState so someone can't setup nodes to counter-DoS based on orphan
-                    // resolution (that is, feeding people an invalid transaction based on LegitTxX in order to get
-                    // anyone relaying LegitTxX banned)
-                    CValidationState stateDummy;
-
-                    if (setMisbehaving.count(mi->fromPeer))
-                        continue;
-                    if (AcceptToMemoryPool(mempool, stateDummy, orphanTx, true, &fMissingInputs2)) {
-                        logDebug(Log::Mempool) << "   accepted orphan tx" << orphanHash;
-                        RelayTransaction(orphanTx);
-                        vWorkQueue.push_back(orphanHash);
-                        vEraseQueue.push_back(orphanHash);
-                    }
-                    else if (!fMissingInputs2)
-                    {
-                        int nDos = 0;
-                        if (stateDummy.IsInvalid(nDos) && nDos > 0)
-                        {
-                            // Punish peer that gave us an invalid orphan tx
-                            Misbehaving(mi->fromPeer, nDos);
-                            setMisbehaving.insert(mi->fromPeer);
-                            logDebug(Log::Mempool) << "   invalid orphan tx" << orphanHash;
-                        }
-                        // Has inputs but not accepted to mempool
-                        // Probably non-standard or insufficient fee/priority
-                        logDebug(Log::Mempool) << "   removed orphan tx" << orphanHash;
-                        vEraseQueue.push_back(orphanHash);
-                        assert(recentRejects);
-                        recentRejects->insert(orphanHash);
-                    }
-                    mempool.check(pcoinsTip);
-                }
-            }
-
-            CTxOrphanCache::instance()->EraseOrphans(vEraseQueue);
-            CTxOrphanCache::instance()->EraseOrphansByTime();
-        }
-        else if (fMissingInputs)
-        {
-            CTxOrphanCache *cache = CTxOrphanCache::instance();
-            // DoS prevention: do not allow CTxOrphanCache to grow unbounded
-            cache->AddOrphanTx(tx, pfrom->GetId());
-            std::uint32_t nEvicted = cache->LimitOrphanTxSize();
-            if (nEvicted > 0)
-                logDebug(Log::Mempool) << "mapOrphan overflow, removed" << nEvicted << "tx";
-        } else {
-            assert(recentRejects);
-            recentRejects->insert(tx.GetHash());
-
-            if (pfrom->fWhitelisted && GetBoolArg("-whitelistforcerelay", DEFAULT_WHITELISTFORCERELAY)) {
-                // Always relay transactions received from whitelisted peers, even
-                // if they were already in the mempool or rejected from it due
-                // to policy, allowing the node to function as a gateway for
-                // nodes hidden behind it.
-                //
-                // Never relay transactions that we would assign a non-zero DoS
-                // score for, as we expect peers to do the same with us in that
-                // case.
-                int nDoS = 0;
-                if (!state.IsInvalid(nDoS) || nDoS == 0) {
-                    LogPrintf("Force relaying tx %s from whitelisted peer=%d\n", tx.GetHash().ToString(), pfrom->id);
-                    RelayTransaction(tx);
-                } else {
-                    LogPrintf("Not relaying invalid transaction %s from whitelisted peer=%d (%s)\n", tx.GetHash().ToString(), pfrom->id, FormatStateMessage(state));
-                }
-            }
-        }
-        int nDoS = 0;
-        if (state.IsInvalid(nDoS)) {
-            logWarning(Log::Mempool) << tx.GetHash() << "from peer" << pfrom->id <<"was not accepted"
-                                     << FormatStateMessage(state);
-            if (state.GetRejectCode() < REJECT_INTERNAL) // Never send AcceptToMemoryPool's internal codes over P2P
-                pfrom->PushMessage(NetMsgType::REJECT, strCommand, (unsigned char)state.GetRejectCode(),
-                                   state.GetRejectReason().substr(0, MAX_REJECT_MESSAGE_LENGTH), inv.hash);
-            if (nDoS > 0)
-                Misbehaving(pfrom->GetId(), nDoS);
-        }
-        FlushStateToDisk(state, FLUSH_STATE_PERIODIC);
-
-        return true;
-    }
-};
-
-class HeadersMessage : public NetworkMessage {
-public:
-    ~HeadersMessage() {}
-
-    HeadersMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        std::vector<CBlockHeader> headers;
-
-        // Bypass the normal CBlock deserialization, as we don't want to risk deserializing 2000 full blocks.
-        unsigned int nCount = ReadCompactSize(vRecv);
-        if (nCount > MAX_HEADERS_RESULTS) {
-            Misbehaving(pfrom->GetId(), 20);
-            return error("headers message size = %u", nCount);
-        }
-        headers.resize(nCount);
-        for (unsigned int n = 0; n < nCount; n++) {
-            vRecv >> headers[n];
-            ReadCompactSize(vRecv); // ignore tx count; assume it is 0.
-        }
-
-        LOCK(cs_main);
-
-        if (nCount == 0) {
-            // Nothing interesting. Stop asking this peers for more headers.
-            return true;
-        }
-
-        CBlockIndex *pindexLast = NULL;
-        BOOST_FOREACH(const CBlockHeader& header, headers) {
-            CValidationState state;
-            if (pindexLast != NULL && header.hashPrevBlock != pindexLast->GetBlockHash()) {
-                Misbehaving(pfrom->GetId(), 20);
-                logWarning(Log::Net) << "p2p HEADERS command received non-continues headers sequence from" << pfrom->GetId();
-                return false;
-            }
-            if (!AcceptBlockHeader(header, state, chainparams, &pindexLast)) {
-                int nDoS;
-                if (state.IsInvalid(nDoS)) {
-                    if (nDoS > 0)
-                        Misbehaving(pfrom->GetId(), nDoS);
-                    logWarning(Log::Net) << "p2p HEADERS command received invalid header from" << pfrom->GetId();
-                    return false;
-                }
-            }
-        }
-
-        if (pindexLast)
-            UpdateBlockAvailability(pfrom->GetId(), pindexLast->GetBlockHash());
-
-        if (nCount == MAX_HEADERS_RESULTS && pindexLast) {
-            // Headers message had its maximum size; the peer may have more headers.
-            // TODO: optimize: if pindexLast is an ancestor of chainActive.Tip or pindexBestHeader, continue
-            // from there instead.
-            logDebug(Log::Net).nospace() << "more getheaders (" << pindexLast->nHeight
-                                         << ") to end to peer=" << pfrom->id << "(startheight:" << pfrom->nStartingHeight << ")";
-            pfrom->PushMessage(NetMsgType::GETHEADERS, chainActive.GetLocator(pindexLast), uint256());
-        }
-
-        if (pindexLast == nullptr) // should never hit.
-            return false;
-
-        bool fCanDirectFetch = CanDirectFetch(chainparams.GetConsensus());
-        CNodeState *nodestate = State(pfrom->GetId());
-        // If this set of headers is valid and ends in a block with at least as
-        // much work as our tip, download as much as possible.
-        if (fCanDirectFetch && pindexLast->IsValid(BLOCK_VALID_TREE) && chainActive.Tip()->nChainWork <= pindexLast->nChainWork) {
-            std::vector<CBlockIndex *> vToFetch;
-            CBlockIndex *pindexWalk = pindexLast;
-            // Calculate all the blocks we'd need to switch to pindexLast, up to a limit.
-            while (pindexWalk && !chainActive.Contains(pindexWalk) && vToFetch.size() <= MAX_BLOCKS_IN_TRANSIT_PER_PEER) {
-                if (!(pindexWalk->nStatus & BLOCK_HAVE_DATA) &&
-                    !mapBlocksInFlight.count(pindexWalk->GetBlockHash())) {
-                    // We don't have this block, and it's not yet in flight.
-                    vToFetch.push_back(pindexWalk);
-                }
-                pindexWalk = pindexWalk->pprev;
-            }
-            // If pindexWalk still isn't on our main chain, we're looking at a
-            // very large reorg at a time we think we're close to caught up to
-            // the main chain -- this shouldn't really happen.  Bail out on the
-            // direct fetch and rely on parallel download instead.
-            if (!chainActive.Contains(pindexWalk)) {
-                LogPrint("net", "Large reorg, won't direct fetch to %s (%d)\n",
-                         pindexLast->GetBlockHash().ToString(),
-                         pindexLast->nHeight);
-            } else if (!xthinEnabled) { // We don't yet support headers-first for XThinblocks.
-                std::vector<CInv> vGetData;
-                // Download as much as possible, from earliest to latest.
-                BOOST_REVERSE_FOREACH(CBlockIndex *pindex, vToFetch) {
-                                if (nodestate->nBlocksInFlight >= MAX_BLOCKS_IN_TRANSIT_PER_PEER) {
-                                    // Can't download any more from this peer
-                                    break;
-                                }
-                                vGetData.push_back(CInv(MSG_BLOCK, pindex->GetBlockHash()));
-                                MarkBlockAsInFlight(pfrom->GetId(), pindex->GetBlockHash(), chainparams.GetConsensus(), pindex);
-                                logDebug(Log::Net) << "Requesting block" << pindex->GetBlockHash() << "from  peer:" << pfrom->id;
-                            }
-                if (vGetData.size() > 1) {
-                    logDebug(Log::Net) << "Downloading blocks toward" << pindexLast->GetBlockHash() << "height:" << pindexLast->nHeight;
-                }
-                if (vGetData.size() > 0) {
-                    pfrom->PushMessage(NetMsgType::GETDATA, vGetData);
-                }
-            }
-        }
-
-        CheckBlockIndex();
-        return true;
-    }
-};
-
-class GetXThinMessage : public NetworkMessage {
-public:
-    ~GetXThinMessage() {}
-
-    GetXThinMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!xthinEnabled) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        CBloomFilter filterMemPool;
-        CInv inv;
-        vRecv >> inv >> filterMemPool;
-        if (inv.type != MSG_XTHINBLOCK && inv.type != MSG_THINBLOCK) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 20);
-            return false;
-        }
-
-        LoadFilter(pfrom, &filterMemPool);
-        pfrom->vRecvGetData.insert(pfrom->vRecvGetData.end(), inv);
-        ProcessGetData(pfrom, chainparams.GetConsensus());
-    }
-};
-
-class XThinBlockMessage : public NetworkMessage {
-public:
-    ~XThinBlockMessage() {}
-
-    XThinBlockMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!xthinEnabled) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        CXThinBlock thinBlock;
-        vRecv >> thinBlock;
-
-        // Send expedited ASAP
-        CValidationState state;
-        if (!CheckBlockHeader(thinBlock.header, state, true)) { // block header is bad
-            LogPrint("thin", "Thinblock %s received with bad header from peer %s (%d)\n", thinBlock.header.GetHash().ToString(), pfrom->addrName.c_str(), pfrom->id);
-            Misbehaving(pfrom->id, 20);
-            return false;
-        }
-        else if (!IsRecentlyExpeditedAndStore(thinBlock.header.GetHash()))
-            SendExpeditedBlock(thinBlock, 0, pfrom);
-
-        CInv inv(MSG_BLOCK, thinBlock.header.GetHash());
-#ifdef LOG_XTHINBLOCKS
-        int nSizeThinBlock = ::GetSerializeSize(thinBlock, SER_NETWORK, PROTOCOL_VERSION);
-        LogPrint("thin", "Received thinblock %s from peer %s (%d). Size %d bytes.\n", inv.hash.ToString(), pfrom->addrName.c_str(), pfrom->id, nSizeThinBlock);
-#endif
-
-        bool fAlreadyHave = false;
-        // An expedited block or re-requested xthin can arrive and beat the original thin block request/response
-        if (!pfrom->mapThinBlocksInFlight.count(inv.hash)) {
-            LogPrint("thin", "Thinblock %s from peer %s (%d) received but we already have it\n", inv.hash.ToString(), pfrom->addrName.c_str(), pfrom->id);
-            LOCK(cs_main);
-            fAlreadyHave = AlreadyHave(inv); // I'll still continue processing if we don't have an accepted block yet
-        }
-
-        if (!fAlreadyHave) {
-            if (thinBlock.process(pfrom))
-                HandleBlockMessage(pfrom, strCommand, pfrom->thinBlock,  thinBlock.GetInv());  // clears the thin block
-        }
-    }
-};
-
-class XBlockTxMessage : public NetworkMessage {
-public:
-    ~XBlockTxMessage() {}
-
-    XBlockTxMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!xthinEnabled) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        if (pfrom->xThinBlockHashes.size() != pfrom->thinBlock.vtx.size()) { // crappy, but fast solution.
-            LogPrint("thin", "Inconsistent thin block data while processing xblock-tx\n");
-            return true;
-        }
-
-        CXThinBlockTx thinBlockTx;
-        vRecv >> thinBlockTx;
-
-        CInv inv(MSG_XTHINBLOCK, thinBlockTx.blockhash);
-        logDebug(Log::Net) << "received blocktxs for" << inv.hash << "peer" << pfrom->id;
-        if (!pfrom->mapThinBlocksInFlight.count(inv.hash)) {
-            LogPrint("thin", "ThinblockTx received but it was either not requested or it was beaten by another block %s  peer=%d\n", inv.hash.ToString(), pfrom->id);
-            return true;
-        }
-
-        // Create the mapMissingTx from all the supplied tx's in the xthinblock
-        std::map<uint64_t, CTransaction> mapMissingTx;
-        BOOST_FOREACH(CTransaction tx, thinBlockTx.vMissingTx) {
-                        mapMissingTx[tx.GetHash().GetCheapHash()] = tx;
-                    }
-
-        int count=0;
-        for (size_t i = 0; i < pfrom->thinBlock.vtx.size(); ++i) {
-            if (pfrom->thinBlock.vtx[i].IsNull()) {
-                auto val = mapMissingTx.find(pfrom->xThinBlockHashes[i]);
-                if (val != mapMissingTx.end()) {
-                    pfrom->thinBlock.vtx[i] = val->second;
-                    --pfrom->thinBlockWaitingForTxns;
-                }
-                count++;
-            }
-        }
-        LogPrint("thin", "Got %d Re-requested txs, needed %d of them\n", thinBlockTx.vMissingTx.size(), count);
-
-        if (pfrom->thinBlockWaitingForTxns == 0) {
-            // We have all the transactions now that are in this block: try to reassemble and process.
-            pfrom->thinBlockWaitingForTxns = -1;
-            pfrom->AddInventoryKnown(inv);
-
-#ifdef LOG_XTHINBLOCKS
-            // for compression statistics, we have to add up the size of xthinblock and the re-requested thinBlockTx.
-            int nSizeThinBlockTx = ::GetSerializeSize(thinBlockTx, SER_NETWORK, PROTOCOL_VERSION);
-            int blockSize = pfrom->thinBlock.GetSerializeSize(SER_NETWORK, CBlock::CURRENT_VERSION);
-            LogPrint("thin", "Reassembled thin block for %s (%d bytes). Message was %d bytes (thinblock) and %d bytes (re-requested tx), compression ratio %3.2f\n",
-                     pfrom->thinBlock.GetHash().ToString(),
-                     blockSize,
-                     pfrom->nSizeThinBlock,
-                     nSizeThinBlockTx,
-                     ((float) blockSize) / ( (float) pfrom->nSizeThinBlock + (float) nSizeThinBlockTx )
-                     );
-#endif
-
-            // For correctness sake, assume all came from the orphans cache
-            std::vector<uint256> orphans;
-            orphans.reserve(pfrom->thinBlock.vtx.size());
-            for (unsigned int i = 0; i < pfrom->thinBlock.vtx.size(); i++) {
-                orphans.push_back(pfrom->thinBlock.vtx[i].GetHash());
-            }
-            HandleBlockMessage(pfrom, strCommand, pfrom->thinBlock, inv);
-            CTxOrphanCache::instance()->EraseOrphans(orphans);
-        }
-        else {
-            LogPrint("thin", "Failed to retrieve all transactions for block\n");
-        }
-        return true;
-    }
-};
-
-class GetXBlockTxMessage : public NetworkMessage {
-public:
-    ~GetXBlockTxMessage() {}
-
-    GetXBlockTxMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!xthinEnabled) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        CXRequestThinBlockTx thinRequestBlockTx;
-        vRecv >> thinRequestBlockTx;
-
-        if (thinRequestBlockTx.setCheapHashesToRequest.empty()) { // empty request??
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        // We use MSG_TX here even though we refer to blockhash because we need to track
-        // how many xblocktx requests we make in case of DOS
-        CInv inv(MSG_TX, thinRequestBlockTx.blockhash);
-        LogPrint("thin", "received get_xblocktx for %s peer=%d\n", inv.hash.ToString(), pfrom->id);
-
-        // Check for Misbehaving and DOS
-        // If they make more than 20 requests in 10 minutes then disconnect them
-        {
-            const uint64_t nNow = GetTime();
-            if (pfrom->nGetXBlockTxLastTime <= 0)
-                pfrom->nGetXBlockTxLastTime = nNow;
-            pfrom->nGetXBlockTxCount *= pow(1.0 - 1.0/600.0, (double)(nNow - pfrom->nGetXBlockTxLastTime));
-            pfrom->nGetXBlockTxLastTime = nNow;
-            pfrom->nGetXBlockTxCount += 1;
-            LogPrint("thin", "nGetXBlockTxCount is %f\n", pfrom->nGetXBlockTxCount);
-            if (pfrom->nGetXBlockTxCount >= 20) {
-                LogPrintf("DOS: Misbehaving - requesting too many xblocktx: %s\n", inv.hash.ToString());
-                LOCK(cs_main);
-                Misbehaving(pfrom->GetId(), 100);  // If they exceed the limit then disconnect them
-            }
-        }
-
-        LOCK(cs_main);
-        auto mi = Blocks::indexMap.find(inv.hash);
-        if (mi == Blocks::indexMap.end()) {
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        CBlockIndex *index = mi->second;
-        assert(index);
-        if (index->nHeight + 100 < chainActive.Height()) {
-            // a node that is behind should never use this method.
-            Misbehaving(pfrom->GetId(), 10);
-            return false;
-        }
-        if ((index->nStatus & BLOCK_HAVE_DATA) == 0) {
-            LogPrintf("GET_XBLOCKTX requested block-data not available %s\n", inv.hash.ToString().c_str());
-            return false;
-        }
-        CBlock block;
-        const Consensus::Params& consensusParams = Params().GetConsensus();
-        if (!ReadBlockFromDisk(block, index, consensusParams)) {
-            LogPrintf("Internal error, file missing datafile %d (block: %d)\n", index->nFile, index->nHeight);
-            return false;
-        }
-
-        std::vector<CTransaction> vTx;
-        int todo = thinRequestBlockTx.setCheapHashesToRequest.size();
-        for (size_t i = 1; i < block.vtx.size(); i++) {
-            uint64_t cheapHash = block.vtx[i].GetHash().GetCheapHash();
-            if (thinRequestBlockTx.setCheapHashesToRequest.count(cheapHash)) {
-                vTx.push_back(block.vtx[i]);
-                if (--todo == 0)
-                    break;
-            }
-        }
-        if (todo > 0) { // node send us a request for transactions which were not in the block.
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-
-        pfrom->AddInventoryKnown(inv);
-        CXThinBlockTx thinBlockTx(thinRequestBlockTx.blockhash, vTx);
-        pfrom->PushMessage(NetMsgType::XBLOCKTX, thinBlockTx);
-
-        return true;
-    }
-};
-
-class BlockMessage : public NetworkMessage {
-public:
-    ~BlockMessage() {}
-
-    BlockMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing) {
-        CBlock block;
-        try {
-            vRecv >> block;
-        } catch (std::exception &e) {
-            LogPrint("net", "ProcessMessage/block failed to parse message and got error: %s\n", e.what());
-            pfrom->fDisconnect = true;
-            return true;
-        }
-
-        CInv inv(MSG_BLOCK, block.GetHash());
-        logDebug(Log::Net) << "received block" << inv << "peer" << pfrom->id;
-
-        pfrom->AddInventoryKnown(inv);
-
-        // BUIP010 Extreme Thinblocks: Handle Block Message
-        HandleBlockMessage(pfrom, strCommand, block, inv);
-        std::vector<uint256> orphans;
-        orphans.reserve(block.vtx.size());
-        for (unsigned int i = 0; i < block.vtx.size(); i++) {
-            orphans.push_back(block.vtx[i].GetHash());
-        }
-        CTxOrphanCache::instance()->EraseOrphans(orphans);
-        return true;
-    }
-};
-
-class GetAddrMessage : public NetworkMessage {
-public:
-    ~GetAddrMessage() {}
-
-    GetAddrMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        // This asymmetric behavior for inbound and outbound connections was introduced
-        // to prevent a fingerprinting attack: an attacker can send specific fake addresses
-        // to users' AddrMan and later request them by sending getaddr messages.
-        // Making nodes which are behind NAT and can only make outgoing connections ignore
-        // the getaddr message mitigates the attack.
-        if (!pfrom->fInbound) {
-            LogPrint("net", "Ignoring \"getaddr\" from outbound connection. peer=%d\n", pfrom->id);
-            return true;
-        }
-
-        // Only send one GetAddr response per connection to reduce resource waste
-        //  and discourage addr stamping of INV announcements.
-        if (pfrom->fSentAddr) {
-            LogPrint("net", "Ignoring repeated \"getaddr\". peer=%d\n", pfrom->id);
-            return true;
-        }
-        pfrom->fSentAddr = true;
-
-        pfrom->vAddrToSend.clear();
-        std::vector<CAddress> vAddr = addrman.GetAddr();
-        BOOST_FOREACH(const CAddress &addr, vAddr)
-                        pfrom->PushAddress(addr);
-        return true;
-    }
-};
-
-class MemPoolMessage : public NetworkMessage {
-public:
-    ~MemPoolMessage() {}
-
-    MemPoolMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (CNode::OutboundTargetReached(false) && !pfrom->fWhitelisted)
-        {
-            LogPrint("net", "mempool request with bandwidth limit reached, disconnect peer=%d\n", pfrom->GetId());
-            pfrom->fDisconnect = true;
-            return true;
-        }
-        LOCK2(cs_main, pfrom->cs_filter);
-
-        std::vector<uint256> vtxid;
-        mempool.queryHashes(vtxid);
-        std::vector<CInv> vInv;
-        BOOST_FOREACH(uint256& hash, vtxid) {
-                        CInv inv(MSG_TX, hash);
-                        if (pfrom->pfilter) {
-                            CTransaction tx;
-                            bool fInMemPool = mempool.lookup(hash, tx);
-                            if (!fInMemPool) continue; // another thread removed since queryHashes, maybe...
-                            if (!pfrom->pfilter->IsRelevantAndUpdate(tx)) continue;
-                        }
-                        vInv.push_back(inv);
-                        if (vInv.size() == MAX_INV_SZ) {
-                            pfrom->PushMessage(NetMsgType::INV, vInv);
-                            vInv.clear();
-                        }
-                    }
-        if (vInv.size() > 0)
-            pfrom->PushMessage(NetMsgType::INV, vInv);
-
-        return true;
-    }
-};
-
-class PingMessage : public NetworkMessage {
-public:
-    ~PingMessage() {}
-
-    PingMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (pfrom->nVersion > BIP0031_VERSION)
-        {
-            uint64_t nonce = 0;
-            vRecv >> nonce;
-            // Echo the message back with the nonce. This allows for two useful features:
-            //
-            // 1) A remote node can quickly check if the connection is operational
-            // 2) Remote nodes can measure the latency of the network thread. If this node
-            //    is overloaded it won't respond to pings quickly and the remote node can
-            //    avoid sending us more work, like chain download requests.
-            //
-            // The nonce stops the remote getting confused between different pings: without
-            // it, if the remote node sends a ping once per second and this node takes 5
-            // seconds to respond to each, the 5th ping the remote sends would appear to
-            // return very quickly.
-            pfrom->PushMessage(NetMsgType::PONG, nonce);
-        }
-        return true;
-    }
-};
-
-class PongMessage : public NetworkMessage {
-public:
-    ~PongMessage() {}
-
-    PongMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        int64_t pingUsecEnd = nTimeReceived;
-        uint64_t nonce = 0;
-        size_t nAvail = vRecv.in_avail();
-        bool bPingFinished = false;
-        std::string sProblem;
-
-        if (nAvail >= sizeof(nonce)) {
-            vRecv >> nonce;
-
-            // Only process pong message if there is an outstanding ping (old ping without nonce should never pong)
-            if (pfrom->nPingNonceSent != 0) {
-                if (nonce == pfrom->nPingNonceSent) {
-                    // Matching pong received, this ping is no longer outstanding
-                    bPingFinished = true;
-                    int64_t pingUsecTime = pingUsecEnd - pfrom->nPingUsecStart;
-                    if (pingUsecTime > 0) {
-                        // Successful ping time measurement, replace previous
-                        pfrom->nPingUsecTime = pingUsecTime;
-                        pfrom->nMinPingUsecTime = std::min(pfrom->nMinPingUsecTime, pingUsecTime);
-                    } else {
-                        // This should never happen
-                        sProblem = "Timing mishap";
-                    }
-                } else {
-                    // Nonce mismatches are normal when pings are overlapping
-                    sProblem = "Nonce mismatch";
-                    if (nonce == 0) {
-                        // This is most likely a bug in another implementation somewhere; cancel this ping
-                        bPingFinished = true;
-                        sProblem = "Nonce zero";
-                    }
-                }
-            } else {
-                sProblem = "Unsolicited pong without ping";
-            }
-        } else {
-            // This is most likely a bug in another implementation somewhere; cancel this ping
-            bPingFinished = true;
-            sProblem = "Short payload";
-        }
-
-        if (!(sProblem.empty())) {
-            LogPrint("net", "pong peer=%d: %s, %x expected, %x received, %u bytes\n",
-                     pfrom->id,
-                     sProblem,
-                     pfrom->nPingNonceSent,
-                     nonce,
-                     nAvail);
-        }
-        if (bPingFinished) {
-            pfrom->nPingNonceSent = 0;
-        }
-
-        return true;
-    }
-};
-
-class FilterLoadMessage : public NetworkMessage {
-public:
-    ~FilterLoadMessage() {}
-
-    FilterLoadMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!GetBoolArg("-peerbloomfilters", true)) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-
-        CBloomFilter filter;
-        vRecv >> filter;
-
-        if (!filter.IsWithinSizeConstraints()) {
-            // There is no excuse for sending a too-large filter
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        } else {
-            LOCK(pfrom->cs_filter);
-            delete pfrom->pfilter;
-            pfrom->pfilter = new CBloomFilter(filter);
-            pfrom->pfilter->UpdateEmptyFull();
-        }
-        pfrom->fRelayTxes = true;
-
-        return true;
-    }
-};
-
-class FilterAddMessage : public NetworkMessage {
-public:
-    ~FilterAddMessage() {}
-
-    FilterAddMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!GetBoolArg("-peerbloomfilters", true)) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        std::vector<unsigned char> vData;
-        vRecv >> vData;
-
-        // Nodes must NEVER send a data item > 520 bytes (the max size for a script data object,
-        // and thus, the maximum size any matched object can have) in a filteradd message
-        if (vData.size() > MAX_SCRIPT_ELEMENT_SIZE) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        } else {
-            LOCK(pfrom->cs_filter);
-            if (pfrom->pfilter)
-                pfrom->pfilter->insert(vData);
-            else
-                Misbehaving(pfrom->GetId(), 100);
-        }
-        return true;
-    }
-};
-
-class FilterClearMessage : public NetworkMessage {
-public:
-    ~FilterClearMessage() {}
-
-    FilterClearMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (!GetBoolArg("-peerbloomfilters", true)) {
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 100);
-            return false;
-        }
-        LOCK(pfrom->cs_filter);
-        delete pfrom->pfilter;
-        pfrom->pfilter = new CBloomFilter();
-        pfrom->fRelayTxes = true;
-        return true;
-    }
-};
-
-class RejectMessage : public NetworkMessage {
-public:
-    ~RejectMessage() {}
-
-    RejectMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        if (fDebug) {
-            try {
-                std::string strMsg; unsigned char ccode; std::string strReason;
-                vRecv >> LIMITED_STRING(strMsg, CMessageHeader::COMMAND_SIZE) >> ccode >> LIMITED_STRING(strReason, MAX_REJECT_MESSAGE_LENGTH);
-
-                std::ostringstream ss;
-                ss << strMsg << " code " << itostr(ccode) << ": " << strReason;
-
-                if (strMsg == NetMsgType::BLOCK || strMsg == NetMsgType::TX)
-                {
-                    uint256 hash;
-                    vRecv >> hash;
-                    ss << ": hash " << hash.ToString();
-                }
-                LogPrint("net", "Reject %s\n", SanitizeString(ss.str()));
-            } catch (const std::ios_base::failure&) {
-                // Avoid feedback loops by preventing reject messages from triggering a new reject message.
-                LogPrint("net", "Unparseable reject message received\n");
-            }
-        }
-        return true;
-    }
-};
-
-class ExpeditedRequestMessage : public NetworkMessage {
-public:
-    ~ExpeditedRequestMessage() {}
-
-    ExpeditedRequestMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        HandleExpeditedRequest(vRecv, pfrom);
-        return true;
-    }
-};
-
-class ExpeditedBlockMessage : public NetworkMessage {
-public:
-    ~ExpeditedBlockMessage() {}
-
-    ExpeditedBlockMessage() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        HandleExpeditedBlock(vRecv, pfrom);
-        return true;
-    }
-};
-
-class Version2Message : public NetworkMessage {
-public:
-    ~Version2Message() {}
-
-    Version2Message() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        // Each connection can only send one version message
-        if (pfrom->addrFromPort != 0) {
-            pfrom->PushMessage(NetMsgType::REJECT, strCommand, REJECT_DUPLICATE,
-                               std::string("Duplicate version2 message"));
-            LOCK(cs_main);
-            Misbehaving(pfrom->GetId(), 15);
-            return false;
-        }
-
-        vRecv >> pfrom->addrFromPort;
-        pfrom->PushMessage(NetMsgType::VERACK2);
-        return true;
-    }
-};
-
-class VersionAck2Message : public NetworkMessage {
-public:
-    ~VersionAck2Message() {}
-
-    VersionAck2Message() : NetworkMessage() {}
-
-    bool handle(CNode * const pfrom,
-                CDataStream &vRecv,
-                int64_t nTimeReceived,
-                std::string &strCommand,
-                const CChainParams &chainparams,
-                const bool xthinEnabled,
-                const bool fReindexing)
-    {
-        CheckAndRequestExpeditedBlocks(pfrom);
-        return true;
-    }
-};
-
 bool static ProcessMessage(CNode* pfrom, std::string strCommand, CDataStream& vRecv, int64_t nTimeReceived)
 {
     const CChainParams& chainparams = Params();
@@ -5802,7 +3809,7 @@ bool static ProcessMessage(CNode* pfrom, std::string strCommand, CDataStream& vR
     }
 
     // FIX-ME: Perhaps a try/catch here according to Ticket 224 description by @zander
-    return networkMessage->handle(pfrom, vRecv, nTimeReceived, strCommand, chainparams, xthinEnabled, fReindex);
+    return networkMessage->handle(pfrom, vRecv, nTimeReceived, strCommand, xthinEnabled, fReindex);
 }
 
 // requires LOCK(cs_vRecvMsg)
