+++ src/messages/AddrMessage.cpp
@@ -1,0 +1,95 @@
+/*
+ * This file is part of the bitcoin-classic project
+ * Copyright (c) 2009-2010 Satoshi Nakamoto
+ * Copyright (c) 2009-2015 The Bitcoin Core developers
+ * Copyright (C) 2017 Tom Zander <tomz@freedommail.ch>
+ * Copyright (C) 2017 Angel Leon <@gubatron>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "AddrMessage.h"
+
+namespace {
+    bool AddrMessage::handle(CNode *const pfrom,
+                             CDataStream &vRecv,
+                             int64_t nTimeReceived,
+                             std::string &strCommand,
+                             const bool xthinEnabled,
+                             const bool fReindexing)
+    {
+        std::vector<CAddress> vAddr;
+        vRecv >> vAddr;
+
+        // Don't want addr from older versions unless seeding
+        if (pfrom->nVersion < CADDR_TIME_VERSION && addrman.size() > 1000)
+            return true;
+        if (vAddr.size() > 1000) {
+            Misbehaving(pfrom->GetId(), 20);
+            return error("message addr size() = %u", vAddr.size());
+        }
+
+        // Store the new addresses
+        std::vector<CAddress> vAddrOk;
+        int64_t nNow = GetAdjustedTime();
+        int64_t nSince = nNow - 10 * 60;
+        BOOST_FOREACH(CAddress &addr, vAddr) {
+                        boost::this_thread::interruption_point();
+
+                        if (addr.nTime <= 100000000 || addr.nTime > nNow + 10 * 60)
+                            addr.nTime = nNow - 5 * 24 * 60 * 60;
+                        pfrom->AddAddressKnown(addr);
+                        bool fReachable = IsReachable(addr);
+                        if (addr.nTime > nSince && !pfrom->fGetAddr && vAddr.size() <= 10 && addr.IsRoutable()) {
+                            // Relay to a limited number of other nodes
+                            {
+                                LOCK(cs_vNodes);
+                                // Use deterministic randomness to send to the same nodes for 24 hours
+                                // at a time so the addrKnowns of the chosen nodes prevent repeats
+                                static uint256 hashSalt;
+                                if (hashSalt.IsNull())
+                                    hashSalt = GetRandHash();
+                                uint64_t hashAddr = addr.GetHash();
+                                uint256 hashRand = ArithToUint256(UintToArith256(hashSalt) ^ (hashAddr << 32) ^
+                                                                  ((GetTime() + hashAddr) / (24 * 60 * 60)));
+                                hashRand = Hash(BEGIN(hashRand), END(hashRand));
+                                std::multimap<uint256, CNode *> mapMix;
+                                BOOST_FOREACH(CNode *pnode, vNodes) {
+                                                if (pnode->nVersion < CADDR_TIME_VERSION)
+                                                    continue;
+                                                unsigned int nPointer;
+                                                memcpy(&nPointer, &pnode, sizeof(nPointer));
+                                                uint256 hashKey = ArithToUint256(UintToArith256(hashRand) ^ nPointer);
+                                                hashKey = Hash(BEGIN(hashKey), END(hashKey));
+                                                mapMix.insert(std::make_pair(hashKey, pnode));
+                                            }
+                                int nRelayNodes = fReachable ? 2
+                                                             : 1; // limited relaying of addresses outside our network(s)
+                                for (std::multimap<uint256, CNode *>::iterator mi = mapMix.begin();
+                                     mi != mapMix.end() && nRelayNodes-- > 0; ++mi)
+                                    ((*mi).second)->PushAddress(addr);
+                            }
+                        }
+                        // Do not store addresses outside our network
+                        if (fReachable)
+                            vAddrOk.push_back(addr);
+                    }
+        addrman.Add(vAddrOk, pfrom->addr, 2 * 60 * 60);
+        if (vAddr.size() < 1000)
+            pfrom->fGetAddr = false;
+        if (pfrom->fOneShot)
+            pfrom->fDisconnect = true;
+        return true;
+    }
+}
