+++ src/messages/HeadersMessage.cpp
@@ -1,0 +1,145 @@
+/*
+ * This file is part of the bitcoin-classic project
+ * Copyright (c) 2009-2010 Satoshi Nakamoto
+ * Copyright (c) 2009-2015 The Bitcoin Core developers
+ * Copyright (C) 2017 Tom Zander <tomz@freedommail.ch>
+ * Copyright (C) 2017 Angel Leon <@gubatron>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "HeadersMessage.h"
+
+namespace {
+    bool HeadersMessage::handle(CNode *const pfrom,
+                                CDataStream &vRecv,
+                                int64_t nTimeReceived,
+                                std::string &strCommand,
+                                const bool xthinEnabled,
+                                const bool fReindexing) {
+        std::vector<CBlockHeader> headers;
+
+        // Bypass the normal CBlock deserialization, as we don't want to risk deserializing 2000 full blocks.
+        unsigned int nCount = ReadCompactSize(vRecv);
+        if (nCount > MAX_HEADERS_RESULTS) {
+            Misbehaving(pfrom->GetId(), 20);
+            return error("headers message size = %u", nCount);
+        }
+        headers.resize(nCount);
+        for (unsigned int n = 0; n < nCount; n++) {
+            vRecv >> headers[n];
+            ReadCompactSize(vRecv); // ignore tx count; assume it is 0.
+        }
+
+        LOCK(cs_main);
+
+        if (nCount == 0) {
+            // Nothing interesting. Stop asking this peers for more headers.
+            return true;
+        }
+
+        const CChainParams chainparams = Params();
+        CBlockIndex *pindexLast = NULL;
+        BOOST_FOREACH(const CBlockHeader &header, headers) {
+                        CValidationState state;
+                        if (pindexLast != NULL && header.hashPrevBlock != pindexLast->GetBlockHash()) {
+                            Misbehaving(pfrom->GetId(), 20);
+                            logWarning(Log::Net)
+                                    << "p2p HEADERS command received non-continues headers sequence from"
+                                    << pfrom->GetId();
+                            return false;
+                        }
+                        if (!AcceptBlockHeader(header, state, chainparams, &pindexLast)) {
+                            int nDoS;
+                            if (state.IsInvalid(nDoS)) {
+                                if (nDoS > 0)
+                                    Misbehaving(pfrom->GetId(), nDoS);
+                                logWarning(Log::Net) << "p2p HEADERS command received invalid header from"
+                                                     << pfrom->GetId();
+                                return false;
+                            }
+                        }
+                    }
+
+        if (pindexLast)
+            UpdateBlockAvailability(pfrom->GetId(), pindexLast->GetBlockHash());
+
+        if (nCount == MAX_HEADERS_RESULTS && pindexLast) {
+            // Headers message had its maximum size; the peer may have more headers.
+            // TODO: optimize: if pindexLast is an ancestor of chainActive.Tip or pindexBestHeader, continue
+            // from there instead.
+            logDebug(Log::Net).nospace() << "more getheaders (" << pindexLast->nHeight
+                                         << ") to end to peer=" << pfrom->id << "(startheight:"
+                                         << pfrom->nStartingHeight << ")";
+            pfrom->PushMessage(NetMsgType::GETHEADERS, chainActive.GetLocator(pindexLast), uint256());
+        }
+
+        if (pindexLast == nullptr) // should never hit.
+            return false;
+
+        bool fCanDirectFetch = CanDirectFetch(chainparams.GetConsensus());
+        CNodeState *nodestate = State(pfrom->GetId());
+        // If this set of headers is valid and ends in a block with at least as
+        // much work as our tip, download as much as possible.
+        if (fCanDirectFetch && pindexLast->IsValid(BLOCK_VALID_TREE) &&
+            chainActive.Tip()->nChainWork <= pindexLast->nChainWork) {
+            std::vector<CBlockIndex *> vToFetch;
+            CBlockIndex *pindexWalk = pindexLast;
+            // Calculate all the blocks we'd need to switch to pindexLast, up to a limit.
+            while (pindexWalk && !chainActive.Contains(pindexWalk) &&
+                   vToFetch.size() <= MAX_BLOCKS_IN_TRANSIT_PER_PEER) {
+                if (!(pindexWalk->nStatus & BLOCK_HAVE_DATA) &&
+                    !mapBlocksInFlight.count(pindexWalk->GetBlockHash())) {
+                    // We don't have this block, and it's not yet in flight.
+                    vToFetch.push_back(pindexWalk);
+                }
+                pindexWalk = pindexWalk->pprev;
+            }
+            // If pindexWalk still isn't on our main chain, we're looking at a
+            // very large reorg at a time we think we're close to caught up to
+            // the main chain -- this shouldn't really happen.  Bail out on the
+            // direct fetch and rely on parallel download instead.
+            if (!chainActive.Contains(pindexWalk)) {
+                LogPrint("net", "Large reorg, won't direct fetch to %s (%d)\n",
+                         pindexLast->GetBlockHash().ToString(),
+                         pindexLast->nHeight);
+            } else if (!xthinEnabled) { // We don't yet support headers-first for XThinblocks.
+                const CChainParams chainparams = Params();
+                std::vector<CInv> vGetData;
+                // Download as much as possible, from earliest to latest.
+                BOOST_REVERSE_FOREACH(CBlockIndex *pindex, vToFetch) {
+                                if (nodestate->nBlocksInFlight >= MAX_BLOCKS_IN_TRANSIT_PER_PEER) {
+                                    // Can't download any more from this peer
+                                    break;
+                                }
+                                vGetData.push_back(CInv(MSG_BLOCK, pindex->GetBlockHash()));
+                                MarkBlockAsInFlight(pfrom->GetId(), pindex->GetBlockHash(),
+                                                    chainparams.GetConsensus(), pindex);
+                                logDebug(Log::Net) << "Requesting block" << pindex->GetBlockHash() << "from  peer:"
+                                                   << pfrom->id;
+                            }
+                if (vGetData.size() > 1) {
+                    logDebug(Log::Net) << "Downloading blocks toward" << pindexLast->GetBlockHash() << "height:"
+                                       << pindexLast->nHeight;
+                }
+                if (vGetData.size() > 0) {
+                    pfrom->PushMessage(NetMsgType::GETDATA, vGetData);
+                }
+            }
+        }
+
+        CheckBlockIndex();
+        return true;
+    }
+}
