+++ src/messages/SendHeadersMessage.cpp
@@ -1,0 +1,24 @@
+//
+// Created by gubatron on 10/14/17.
+//
+
+#include "SendHeadersMessage.h"
+
+namespace {
+
+    bool SendHeadersMessage::handle(CNode *const pfrom,
+                                    CDataStream &vRecv,
+                                    int64_t nTimeReceived,
+                                    std::string &strCommand,
+                                    const bool xthinEnabled,
+                                    const bool fReindexing)
+    {
+        LOCK(cs_main);
+        // BUIP010 Xtreme Thinblocks: We only do inv/getdata for xthinblocks and so we must have headersfirst turned off
+        if (xthinEnabled)
+            State(pfrom->GetId())->fPreferHeaders = false;
+        else
+            State(pfrom->GetId())->fPreferHeaders = true;
+        return true;
+    }
+}
