+++ src/messages/SendHeadersMessage.h
@@ -1,0 +1,27 @@
+//
+// Created by gubatron on 10/14/17.
+//
+
+#ifndef BITCOINCLASSIC_SENDHEADERSMESSAGE_H
+#define BITCOINCLASSIC_SENDHEADERSMESSAGE_H
+
+#include "NetworkMessage.h"
+
+namespace {
+    class SendHeadersMessage :
+            public NetworkMessage {
+    public:
+
+        ~SendHeadersMessage() {}
+
+        SendHeadersMessage() : NetworkMessage() {}
+
+        bool handle(CNode *const pfrom,
+                    CDataStream &vRecv,
+                    int64_t nTimeReceived,
+                    std::string &strCommand,
+                    const bool xthinEnabled,
+                    const bool fReindexing);
+    };
+}
+#endif //BITCOINCLASSIC_SENDHEADERSMESSAGE_H
