+++ src/messages/messages.h
@@ -1,0 +1,52 @@
+/*
+ * This file is part of the bitcoin-classic project
+ * Copyright (c) 2009-2010 Satoshi Nakamoto
+ * Copyright (c) 2009-2015 The Bitcoin Core developers
+ * Copyright (C) 2017 Tom Zander <tomz@freedommail.ch>
+ * Copyright (C) 2017 Angel Leon <@gubatron>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef BITCOINCLASSIC_MESSAGES_H
+#define BITCOINCLASSIC_MESSAGES_H
+
+#include "NetworkMessage.h"
+#include "AddrMessage.h"
+#include "BlockMessage.h"
+#include "ExpeditedBlockMessage.h"
+#include "ExpeditedRequestMessage.h"
+#include "FilterAddMessage.h"
+#include "FilterClearMessage.h"
+#include "FilterLoadMessage.h"
+#include "GetBlocksMessage.h"
+#include "GetDataMessage.h"
+#include "GetHeadersMessage.h"
+#include "GetXBlockTxMessage.h"
+#include "GetXThinMessage.h"
+#include "HeadersMessage.h"
+#include "InvMessage.h"
+#include "MemPoolMessage.h"
+#include "SendHeadersMessage.h"
+#include "PingMessage.h"
+#include "PongMessage.h"
+#include "RejectMessage.h"
+#include "TxMessage.h"
+#include "VerAckMessage.h"
+#include "VersionMessage.h"
+#include "XThinBlockMessage.h"
+#include "XBlockTxMessage.h"
+#include "GetAddrMessage.h"
+
+#endif //BITCOINCLASSIC_MESSAGES_H
